der: rename `big_uint` module
From `big_integer` previously.
